pyproject.toml
[project]
name = "tmux-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[cli,mcp,openai]>=1.0.10",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "ruff",
]
